// web/src/app/api/fieldservice.ts
export interface components {
    schemas: {
        Schema: {
            project_slug: string;
            key: string;
            name: string;
            /** Format: int64 */
            created_by: number;
            /** Format: date-time */
            created_at: string;
            /** Format: date-time */
            updated_at: string;
        };
        SchemaWithFields: components["schemas"]["Schema"] & {
            fields?: components["schemas"]["SchemaField"][];
        };
        SchemaField: {
            key: string;
            type: components["schemas"]["FieldType"];
            /** Format: int64 */
            created_by: number;
            /** Format: date-time */
            created_at: string;
            /** Format: date-time */
            updated_at: string;
        };
        /** @enum {string} */
        FieldType: "string" | "int" | "float" | "boolean" | "date" | "datetime";
        CreateSchemaRequest: {
            project_slug: string;
            key: string;
            name: string;
        };
        CreateSchemaFieldRequest: {
            key: string;
            type: components["schemas"]["FieldType"];
        };
        ErrorResponse: {
            error: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}